+++ tests/auth_test.go
@@ -8,7 +8,7 @@ import (
 	"echo-demo-project/internal/responses"
 	"echo-demo-project/server"
 	"echo-demo-project/server/handlers"
-	"echo-demo-project/services/token"
+	"echo-demo-project/internal/services/token"
 	"echo-demo-project/tests/helpers"
 	"encoding/json"
 	"errors"
